+++ src/components/Header.js
@@ -1,0 +1,16 @@
+import React from "react";
+
+const Header = (props) => {
+    return (
+        <div id='header'>
+            <h1>{props.title}</h1>
+            {props.subTitle && <h2>{props.subTitle}</h2>}
+        </div>
+    );
+}
+
+Header.defaultProps = {
+    title: 'Indecision Application'
+}
+
+export default Header;
